requirements.txt
python-dotenv
aiogram==3.3.0
playwright==1.44.0
apscheduler
